+++ src/main/kotlin/graphics/scenery/net/NodePublisher.kt
@@ -7,7 +7,7 @@ import com.esotericsoftware.kryo.io.Output
 import com.jogamp.opengl.math.Quaternion
 import graphics.scenery.*
 import graphics.scenery.utils.Statistics
-import graphics.scenery.volumes.DirectVolumeFullscreen
+import graphics.scenery.volumes.Volume
 import org.slf4j.LoggerFactory
 import org.zeromq.ZContext
 import org.zeromq.ZMQ
@@ -36,7 +36,7 @@ class NodePublisher(override var hub: Hub?, val address: String = "tcp://*:6666"
         kryo.register(DetachedHeadCamera::class.java)
         kryo.register(Quaternion::class.java)
         kryo.register(Mesh::class.java)
-        kryo.register(DirectVolumeFullscreen::class.java)
+        kryo.register(Volume::class.java)
     }
 
     fun publish() {
