+++ src/main/resources/graphics/scenery/backends/vulkan/shaders/Volume.vert
