+++ src/test/tests/graphics/scenery/tests/examples/BileExample.kt
@@ -3,22 +3,10 @@ package graphics.scenery.tests.examples
 import cleargl.GLVector
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
-import graphics.scenery.controls.InputHandler
-import graphics.scenery.controls.OpenVRHMD
 import graphics.scenery.controls.TrackedStereoGlasses
-import graphics.scenery.controls.behaviours.ArcballCameraControl
-import graphics.scenery.controls.behaviours.FPSCameraControl
 import graphics.scenery.net.NodePublisher
 import graphics.scenery.net.NodeSubscriber
-import graphics.scenery.utils.Numerics
-import graphics.scenery.volumes.DirectVolume
-import graphics.scenery.volumes.DirectVolumeFullscreen
-import graphics.scenery.volumes.Volume
 import org.junit.Test
-import org.scijava.ui.behaviour.ClickBehaviour
-import java.io.File
-import java.nio.file.Paths
-import kotlin.concurrent.thread
 
 /**
  * <Description>
