+++ src/test/tests/graphics/scenery/tests/examples/VolumeDrosoExample.kt
@@ -4,14 +4,11 @@ import cleargl.GLVector
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.controls.InputHandler
-import graphics.scenery.controls.OpenVRHMD
 import graphics.scenery.controls.TrackedStereoGlasses
 import graphics.scenery.controls.behaviours.ArcballCameraControl
 import graphics.scenery.controls.behaviours.FPSCameraControl
 import graphics.scenery.net.NodePublisher
 import graphics.scenery.net.NodeSubscriber
-import graphics.scenery.volumes.DirectVolume
-import graphics.scenery.volumes.DirectVolumeFullscreen
 import graphics.scenery.volumes.Volume
 import org.junit.Test
 import org.scijava.ui.behaviour.ClickBehaviour
@@ -52,7 +49,7 @@ class VolumeDrosoExample: SceneryDefaultApplication("Clustered Volume Rendering 
         shell.material.ambient = GLVector.getNullVector(3)
         scene.addChild(shell)
 
-        val volume = DirectVolumeFullscreen(autosetProperties = false)
+        val volume = Volume(autosetProperties = false)
 
         with(volume) {
             trangemin = 0.001f
